web_server_lobby/pyproject.toml
[tool.poetry]
name = "web-server-lobby"
version = "0.1.0"
description = ""
authors = ["AeroX2 <james@ridey.email>"]

[tool.poetry.dependencies]
python = "^3.10"
rq = "^1.11.1"
python-socketio = "^5.7.2"
uvicorn = {extras = ["standard"], version = "^0.19.0"}


[tool.poetry.group.dev.dependencies]
black = "^22.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
